Add block-light channel via a uniform 32-bit voxel word (#28)
* Add block-light channel via a uniform 32-bit voxel word

Voxel light was a single sky-only channel that merely scaled the PBR sun,
so at night (sun off) locally-lit voxels could not glow. Widen the packed
voxel word from 16-bit to 32-bit everywhere (resident memory, mesh worker,
wire, IndexedDB, server/LevelDB) — one consistent Uint32Array format — and
add an independent block-light channel at bit 16 for emitters (e.g. lava).

The low 16 bits stay byte-identical, so every existing pack/unpack shift and
mask is unchanged. Legacy 16-bit persisted records widen losslessly on read
(IndexedDB and LevelDB); light bits are recomputed client-side on ingest.

Lighting: sky light stays sky-column only (emitters no longer merged in);
block light is an emitter-seeded BFS into the new field, cleared and rebuilt
each relight so removing an emitter darkens correctly. LUT indices mask off
the new bits to keep the 2048-entry tables intact.

Shader: sky and block terms are combined by per-channel max, so the sun wins
in daylight (look unchanged) and block light wins at night / in caves. Block
light reaches the GPU as a new per-vertex blockLight attribute; colour and
intensity are live-tunable in the debug panel. New debug modes: SkyLight
(renamed from VoxelLight) and BlockLight.


* Fix block-light removal + Sunlight rename + debug-mode dropdown

Block light removal left stale glow: relightModifiedChunk recomputed the edited
chunk first while neighbours still held its light, and max-merge propagation
re-absorbed that stale light (it can only raise values). It also never covered
diagonal neighbours, which block light reaches (range ~31 voxels ≈ one chunk).

Split block propagation into clearBlockLight + propagateBlockLight (monotonic,
no clear). On edit, relight the 3x3x3 chunk region around the edit: clear the
whole region's block field once, then propagate to a fixed point — so removing
an emitter darkens correctly and diagonal neighbours are covered. Gated on an
emitter being present or existing block light, so ordinary edits stay cheap;
sky light stays a per-chunk recompute (order-independent, anchored by the sun
column) via a new computeSkyLight entry point. Chunk.hasBlockLight caches the
gate. computeAndPropagateLight now returns whether block light is present.

UI: rename the "SkyLight"/"VoxelLight" debug mode to "Sunlight"; make the debug
mode selector a dropdown (faster on mobile) ordered with BlockLight right after
Sunlight via TERRAIN_DEBUG_MODE_ORDER (decoupled from shader mode numbers).


---------

+++ client/src/game/voxel/MeshWorkerPool.ts
@@ -2,7 +2,7 @@
  * MeshWorkerPool - Manages a pool of mesh workers with grid buffer recycling
  * 
  * Dispatches SurfaceNet + geometry expansion work to web workers.
- * Grid buffers (Uint16Array, 34³) are recycled via Transferable round-trips
+ * Grid buffers (Uint32Array, 34³) are recycled via Transferable round-trips
  * to avoid per-chunk allocations.
  * 
  * Two queues: priority (build preview) and regular (chunk remesh).
@@ -21,7 +21,7 @@ const GRID_LENGTH = GRID_SIZE * GRID_SIZE * GRID_SIZE;
 interface Task {
   id: number;
   chunkKey: string;
-  grid: Uint16Array;
+  grid: Uint32Array;
   skipHighBoundary: [boolean, boolean, boolean];
 }
 
@@ -54,7 +54,7 @@ export class MeshWorkerPool {
   private pendingCallbacks = new Map<number, (resp: MeshWorkerResponse) => void>();
 
   /** Pool of reusable grid buffers (returned by workers after each job) */
-  private gridPool: Uint16Array[] = [];
+  private gridPool: Uint32Array[] = [];
 
   /** Track which chunk keys are in-flight (prevent duplicate dispatch) */
   private inFlight = new Set<string>();
@@ -76,7 +76,7 @@ export class MeshWorkerPool {
 
     // Pre-allocate grid buffers (one per worker + a few spare)
     for (let i = 0; i < poolSize + 2; i++) {
-      this.gridPool.push(new Uint16Array(GRID_LENGTH));
+      this.gridPool.push(new Uint32Array(GRID_LENGTH));
     }
   }
 
@@ -84,8 +84,8 @@ export class MeshWorkerPool {
    * Take a grid buffer from the pool (or allocate if empty).
    * Caller fills this, then passes to dispatch().
    */
-  takeGrid(): Uint16Array {
-    return this.gridPool.pop() ?? new Uint16Array(GRID_LENGTH);
+  takeGrid(): Uint32Array {
+    return this.gridPool.pop() ?? new Uint32Array(GRID_LENGTH);
   }
 
   /**
@@ -110,7 +110,7 @@ export class MeshWorkerPool {
    */
   dispatch(
     chunkKey: string,
-    grid: Uint16Array,
+    grid: Uint32Array,
     skipHighBoundary: [boolean, boolean, boolean],
     callback: MeshResultCallback,
   ): void {
@@ -149,7 +149,7 @@ export class MeshWorkerPool {
   dispatchBatch(
     items: Array<{
       chunkKey: string;
-      grid: Uint16Array;
+      grid: Uint32Array;
       skipHighBoundary: [boolean, boolean, boolean];
     }>,
     callback: (results: MeshResult[]) => void,
